* winscp564: (32 commits)
  fix warning: Dereference after null check
  fix warning: Uninitialized scalar field
  update tinyxml2 to latest version
  fix PVS warning: C-style explicit type casting is utilized. Consider using: static_cast/const_cast/reinterpret_cast.
  fix PVS warning: Decreased performance. The 'Value' argument should probably be rendered as a constant reference
  fix cppcheck warning: Clarify calculation precedence for '&' and '?'
  remove unneeded brackets
  fix code style
  update tinyxml2 to latest version
  remove unneeded brackets
  fix code style
  fix cppcheck warning: Clarify calculation precedence for '&' and '?'
  fix x64 build
  version info: update
  update WinSCP version info
  sync with WinSCP 5.6.4 (NetBox)
  sync with WinSCP 5.6.4 (filezilla)
  sync with WinSCP 5.6.4 (resources)
  sync with WinSCP 5.6.4 (Terminal)
  sync with WinSCP 5.6.4 (SftpFileSystem)
  ...

Conflicts:
	libs/tinyxml2/tinyxml2.cpp
	libs/tinyxml2/tinyxml2.h
	src/core/Queue.cpp

+++ src/core/FileBuffer.cpp
@@ -71,7 +71,7 @@ int64_t TFileBuffer::ReadStream(TStream * Stream, const int64_t Len, bool ForceL
     }
     else
     {
-      Result = Stream->Read(GetData() + GetPosition(), (int64_t)Len);
+      Result = Stream->Read(GetData() + GetPosition(), Len);
     }
     if (Result != Len)
     {
@@ -210,13 +210,13 @@ void TFileBuffer::Convert(TEOLType Source, char * Dest, intptr_t Params,
 void TFileBuffer::Insert(int64_t Index, const char * Buf, size_t Len)
 {
   SetSize(GetSize() + Len);
-  memmove(GetData() + Index + Len, GetData() + Index, (size_t)(GetSize() - Index - Len));
+  memmove(GetData() + Index + Len, GetData() + Index, static_cast<size_t>(GetSize() - Index - Len));
   memmove(GetData() + Index, Buf, Len);
 }
 
 void TFileBuffer::Delete(int64_t Index, size_t Len)
 {
-  memmove(GetData() + Index, GetData() + Index + Len, (size_t)(GetSize() - Index - Len));
+  memmove(GetData() + Index, GetData() + Index + Len, static_cast<size_t>(GetSize() - Index - Len));
   SetSize(GetSize() - Len);
 }
